Add printf format checks to the cache manager
Add printf format checking to the afs_warn and afs_warnuser functions
and fix the fallout.

Reviewed-on: http://gerrit.openafs.org/792
Tested-by: Derrick Brashear <shadow@dementia.org>

+++ src/afs/afs_memcache.c
@@ -243,8 +243,8 @@ afs_MemWriteUIO(afs_dcache_id_t *ainode, struct uio *uioP)
 	if (mceP->data == NULL) {	/* no available memory */
 	    mceP->data = oldData;	/* revert back change that was made */
 	    ReleaseWriteLock(&mceP->afs_memLock);
-	    afs_warn("afs: afs_MemWriteBlk mem alloc failure (%d bytes)\n",
-		     AFS_UIO_RESID(uioP) + AFS_UIO_OFFSET(uioP));
+	    afs_warn("afs: afs_MemWriteBlk mem alloc failure (%ld bytes)\n",
+		     (long)(AFS_UIO_RESID(uioP) + AFS_UIO_OFFSET(uioP)));
 	    return -ENOMEM;
 	}
